+++ src/test/java/org/opentripplanner/routing/stoptimes/StopTimesHelperTest.java
@@ -42,11 +42,28 @@ class StopTimesHelperTest {
   }
 
   /**
+   * Case 0, requested number of departure = 0
+   */
+  @Test
+  void stopTimesForStop_zeroRequestedNumberOfDeparture() {
+    var result = StopTimesHelper.stopTimesForStop(
+      transitService,
+      transitService.getRegularStop(stopId),
+      serviceDate.atStartOfDay(transitService.getTimeZone()).toInstant(),
+      Duration.ofHours(24),
+      0,
+      ArrivalDeparture.BOTH,
+      true
+    );
+
+    assertTrue(result.isEmpty());
+  }
+
+  /**
    * Case 1, should find first departure for each pattern when numberOfDepartures is one
    */
   @Test
   void stopTimesForStop_oneDeparture() {
-    // Case 1, should find first departure for each pattern
     var result = StopTimesHelper.stopTimesForStop(
       transitService,
       transitService.getRegularStop(stopId),